TableGenerator/requirements.txt
python-docx>=0.8.11
openai>=0.27.0

